refactor(backend): move messages providers to allow more flexible logging configuration
