+++ backend/infrastructure/src/main/java/com/wilzwert/myjobs/infrastructure/adapter/message/AccountCreationMessageProviderAdapter.java
@@ -1,34 +1,40 @@
-package com.wilzwert.myjobs.infrastructure.adapter;
+package com.wilzwert.myjobs.infrastructure.adapter.message;
 
 import com.wilzwert.myjobs.core.domain.model.user.User;
-import com.wilzwert.myjobs.core.domain.model.user.ports.driven.EmailVerificationMessageProvider;
+import com.wilzwert.myjobs.core.domain.model.user.ports.driven.AccountCreationMessageProvider;
 import com.wilzwert.myjobs.infrastructure.mail.MailProvider;
+import jakarta.mail.MessagingException;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
+import java.io.UnsupportedEncodingException;
+
 @Component
-public class EmailVerificationMessageProviderAdapter implements EmailVerificationMessageProvider {
+@Slf4j
+public class AccountCreationMessageProviderAdapter implements AccountCreationMessageProvider {
 
     private final MailProvider mailProvider;
 
-    public EmailVerificationMessageProviderAdapter(final MailProvider mailProvider) {
+    public AccountCreationMessageProviderAdapter(final MailProvider mailProvider) {
         this.mailProvider = mailProvider;
     }
 
     @Override
     public void send(User user) {
         try {
-            var message = mailProvider.createMessage("mail/email_verification", user.getEmail(), user.getFirstName(), "email.email_verification.subject", user.getLang().toString());
+            var message = mailProvider.createMessage("mail/account_creation", user.getEmail(), user.getFirstName(), "email.account_creation.subject", user.getLang().toString());
 
             message.setVariable("url", mailProvider.createMeUrl(message.getLocale()));
             message.setVariable("firstName", user.getFirstName());
             message.setVariable("lastName", user.getLastName());
             message.setVariable("validationUrl", mailProvider.createUrl("uri.email_validation", message.getLocale(), user.getEmailValidationCode()));
+            log.info("Sending account creation message: {}", user.getEmail());
             mailProvider.send(message);
         }
         // TODO : improve exception handling
-        catch (Exception e) {
+        catch (MessagingException |UnsupportedEncodingException e) {
+            log.info("Sending account creation message failed.", e);
             throw new RuntimeException(e);
         }
-
     }
 }
