+++ backend/infrastructure/src/main/java/com/wilzwert/myjobs/infrastructure/adapter/message/EmailVerificationMessageProviderAdapter.java
@@ -1,37 +1,36 @@
-package com.wilzwert.myjobs.infrastructure.adapter;
+package com.wilzwert.myjobs.infrastructure.adapter.message;
 
 import com.wilzwert.myjobs.core.domain.model.user.User;
-import com.wilzwert.myjobs.core.domain.model.user.ports.driven.AccountCreationMessageProvider;
+import com.wilzwert.myjobs.core.domain.model.user.ports.driven.EmailVerificationMessageProvider;
 import com.wilzwert.myjobs.infrastructure.mail.MailProvider;
-import jakarta.mail.MessagingException;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
-import java.io.UnsupportedEncodingException;
-
 @Component
-public class AccountCreationMessageProviderAdapter implements AccountCreationMessageProvider {
+@Slf4j
+public class EmailVerificationMessageProviderAdapter implements EmailVerificationMessageProvider {
 
     private final MailProvider mailProvider;
 
-    public AccountCreationMessageProviderAdapter(final MailProvider mailProvider) {
+    public EmailVerificationMessageProviderAdapter(final MailProvider mailProvider) {
         this.mailProvider = mailProvider;
     }
 
     @Override
     public void send(User user) {
         try {
-            var message = mailProvider.createMessage("mail/account_creation", user.getEmail(), user.getFirstName(), "email.account_creation.subject", user.getLang().toString());
-
+            var message = mailProvider.createMessage("mail/email_verification", user.getEmail(), user.getFirstName(), "email.email_verification.subject", user.getLang().toString());
             message.setVariable("url", mailProvider.createMeUrl(message.getLocale()));
             message.setVariable("firstName", user.getFirstName());
             message.setVariable("lastName", user.getLastName());
             message.setVariable("validationUrl", mailProvider.createUrl("uri.email_validation", message.getLocale(), user.getEmailValidationCode()));
+            log.info("Sending email verification message: {}", user.getEmail());
             mailProvider.send(message);
         }
         // TODO : improve exception handling
-        catch (MessagingException |UnsupportedEncodingException e) {
+        catch (Exception e) {
+            log.error("Sending email verification message failed.", e);
             throw new RuntimeException(e);
         }
-
     }
 }
