+++ backend/infrastructure/src/test/java/com/wilzwert/myjobs/infrastructure/adapter/JobsReminderMessageProviderAdapterIT.java
@@ -6,6 +6,7 @@ import com.wilzwert.myjobs.core.domain.model.user.EmailStatus;
 import com.wilzwert.myjobs.core.domain.model.user.Lang;
 import com.wilzwert.myjobs.core.domain.model.user.User;
 import com.wilzwert.myjobs.core.domain.model.user.UserId;
+import com.wilzwert.myjobs.infrastructure.adapter.message.JobReminderMessageProviderAdapter;
 import com.wilzwert.myjobs.infrastructure.configuration.AbstractBaseIntegrationTest;
 import com.wilzwert.myjobs.infrastructure.configuration.SyncTestExecutorConfiguration;
 import com.wilzwert.myjobs.infrastructure.utility.EmailUtility;
